+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## 0.0.2-alpha
 Pending release
 
+#### Added
+- Function `getVariable(_:)`.
+
 #### Fixed
 - `Embed.clear()` not removing the timestamp.
 - `PartialInvite.inviter` error when missing inviter.
@@ -8,4 +11,4 @@ Pending release
 ## 0.0.1-alpha
 Released on May 17, 2023.
 
-Discord.swift initial release
+Discord.swift initial release.